HBAPP-95 Move button
