+++ components/features/chat/chat-prompt.tsx
@@ -5,7 +5,7 @@ import { useEnterSubmit } from '@/lib/hooks/use-enter-submit'
 import { Button } from '@/components/ui/button'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip'
 import { IconArrowElbow } from '@/components/ui/icons'
-import { CharmMenuButton } from '@/components/features/menus/charm/button-charm-menu-trigger'
+import { CharmMenuButton } from '@/components/features/chat/button-charm-menu-trigger'
 import { MouseEvent } from 'react'
 
 export interface PromptProps extends Pick<UseChatHelpers, 'input' | 'setInput'> {
